feat(lms): add endpoints for filtered content and user todos
Introduces a new `GET` endpoint for `ModuleContent` to allow
filtering and pagination.
Adds a new `GET /modules/todo` endpoint for students to view their
upcoming assignments and quizzes.
Refactors by moving the `findTodos` method from `LmsContentService`
to `LmsService` for better architectural separation.
Updates the `LmsController` base path from `/lms` to `/modules`
for improved API consistency.
