+++ backend/src/modules/lms/lms-content.service.ts
@@ -33,7 +33,6 @@ import { FileService } from '@/modules/lms/content/file/file.service';
 import { UrlService } from '@/modules/lms/content/url/url.service';
 import { VideoService } from '@/modules/lms/content/video/video.service';
 import { LessonService } from '@/modules/lms/content/lesson/lessson.service';
-import { PaginatedTodosDto } from '@/modules/lms/dto/paginated-todos.dto';
 import { PaginatedModuleContentDto } from '@/modules/lms/dto/paginated-module-content.dto';
 import { FilterModuleContentsDto } from '@/modules/lms/dto/filter-module-contents.dto';
 
@@ -692,153 +691,6 @@ export class LmsContentService {
       });
 
     return { moduleContents: items, meta };
-  }
-
-  @Log({
-    logArgsMessage: ({ userId, page, limit }) =>
-      `Fetching todos for user ${userId} in active term (page: ${page}, limit: ${limit})`,
-    logSuccessMessage: (result) =>
-      `Successfully fetched ${result.todos.length} todos`,
-    logErrorMessage: (err, { userId }) =>
-      `Error fetching todos for user ${userId}: ${err.message}`,
-  })
-  async findTodos(
-    @LogParam('userId') userId: string,
-    @LogParam('page') page: number = 1,
-  ): Promise<PaginatedTodosDto> {
-    if (!isUUID(userId)) {
-      throw new BadRequestException('Invalid user ID format');
-    }
-
-    // First, get the active enrollment period
-    const activeTerm = await this.prisma.client.enrollmentPeriod.findFirst({
-      where: {
-        status: 'active',
-        startDate: { lte: new Date() },
-        endDate: { gte: new Date() },
-      },
-    });
-
-    if (!activeTerm) {
-      return {
-        todos: [],
-        meta: {
-          isFirstPage: true,
-          isLastPage: true,
-          currentPage: page,
-          previousPage: page,
-          nextPage: page,
-          pageCount: 1,
-          totalCount: 0,
-        },
-      };
-    }
-
-    // Get user's enrolled courses in active term
-    const userEnrollments = await this.prisma.client.courseEnrollment.findMany({
-      where: {
-        studentId: userId,
-        status: 'enrolled',
-        courseOffering: {
-          periodId: activeTerm.id,
-        },
-      },
-      include: {
-        courseOffering: {
-          include: {
-            course: true,
-          },
-        },
-      },
-    });
-
-    const courseOfferingIds = userEnrollments.map(
-      (enrollment) => enrollment.courseOfferingId,
-    );
-
-    // Get todos (assignments and quizzes with due dates)
-    const whereCondition: Prisma.ModuleContentWhereInput = {
-      module: {
-        courseOfferingId: { in: courseOfferingIds },
-      },
-      OR: [
-        {
-          contentType: 'ASSIGNMENT',
-          assignment: {
-            dueDate: { gte: new Date() },
-          },
-        },
-        {
-          contentType: 'QUIZ',
-          quiz: {
-            dueDate: { gte: new Date() },
-          },
-        },
-      ],
-      publishedAt: { lte: new Date() },
-    };
-
-    const [todos, meta] = await this.prisma.client.moduleContent
-      .paginate({
-        where: whereCondition,
-        include: {
-          assignment: true,
-          quiz: true,
-          module: {
-            include: {
-              courseOffering: {
-                include: {
-                  course: true,
-                  enrollmentPeriod: true,
-                },
-              },
-            },
-          },
-          studentProgress: {
-            where: { userId },
-          },
-        },
-        orderBy: [
-          {
-            assignment: {
-              dueDate: 'asc',
-            },
-          },
-          {
-            quiz: {
-              dueDate: 'asc',
-            },
-          },
-        ],
-      })
-      .withPages({
-        limit: 10,
-        page,
-        includePageCount: true,
-      });
-
-    // Transform the results to include progress status
-    const items = todos.map((todo) => {
-      const title = todo.assignment?.title || todo.quiz?.title;
-      const dueDate = todo.assignment?.dueDate || todo.quiz?.dueDate;
-
-      if (!title) {
-        throw new Error(`Content ${todo.id} is missing a title`);
-      }
-      if (!dueDate) {
-        throw new Error(`Content ${todo.id} is missing a due date`);
-      }
-
-      return {
-        id: todo.id,
-        type: todo.contentType,
-        title,
-        dueDate,
-        moduleName: todo.module.title,
-      };
-    });
-
-    return { todos: items, meta };
   }
 
   /**
